+++ src/Tests/Analyzers.Tests/RCS1031RemoveUnnecessaryBracesInSwitchSectionTests.cs
@@ -141,6 +141,57 @@ class C
 }
 ");
     }
+    
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBracesInSwitchSection)]
+    public async Task Test_WithLocalVariablesThatDoNotOverlap()
+    {
+        await VerifyDiagnosticAndFixAsync(@"
+using System;
+
+class C
+{
+    void M()
+    {
+        string s = null;
+
+        switch (s)
+        {
+            case """":
+                [|{|]
+                    var x = 1;
+                    break;
+                }
+            default:
+                [|{|]
+                    var y = 1;
+                    break;
+                }
+        }
+    }
+}
+",@"
+using System;
+
+class C
+{
+    void M()
+    {
+        string s = null;
+
+        switch (s)
+        {
+            case """":
+                var x = 1;
+                break;
+
+            default:
+                var y = 1;
+                break;
+        }
+    }
+}
+");
+    }
 
     [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBracesInSwitchSection)]
     public async Task TestNoDiagnostic_SectionWithoutBlock()
@@ -186,6 +237,7 @@ class C
 }
 ");
     }
+    
     [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.RemoveUnnecessaryBracesInSwitchSection)]
     public async Task TestNoDiagnostic_WhenOverlappingLocalVariableDeclaration()
     {